Factor out test django app
Move the test django app out of test_django.py into its own package.
This makes it possible for it to have its own tests.
