+++ appmap/test/data/django/app/hello_world.html
@@ -1,0 +1,3 @@
+<body>
+  <h1>Hello World!</h1>
+</body>
