+++ appmap/test/data/django/appmap.yml
@@ -1,0 +1,4 @@
+name: test_app
+packages:
+- path: app
+
